Move Permutation struct from crate::circuit -> plonk::circuit
