+++ src/circuit/layouter.rs
@@ -5,8 +5,8 @@ use std::collections::{HashMap, HashSet};
 use std::fmt;
 use std::marker::PhantomData;
 
-use super::{Cell, Chip, Layouter, Permutation, Region, RegionIndex, RegionStart};
-use crate::plonk::{Advice, Any, Assignment, Column, Error, Fixed};
+use super::{Cell, Chip, Layouter, Region, RegionIndex, RegionStart};
+use crate::plonk::{Advice, Any, Assignment, Column, Error, Fixed, Permutation};
 
 /// Helper trait for implementing a custom [`Layouter`].
 ///
@@ -321,18 +321,18 @@ impl<'r, 'a, C: Chip, CS: Assignment<C::Field> + 'a> RegionLayouter<C>
         right: Cell,
     ) -> Result<(), Error> {
         let left_column = permutation
-            .mapping
+            .mapping()
             .iter()
             .position(|c| c == &left.column)
             .ok_or(Error::SynthesisError)?;
         let right_column = permutation
-            .mapping
+            .mapping()
             .iter()
             .position(|c| c == &right.column)
             .ok_or(Error::SynthesisError)?;
 
         self.layouter.cs.copy(
-            permutation.index,
+            permutation.index(),
             left_column,
             *self.layouter.regions[*left.region_index] + left.row_offset,
             right_column,
